Clean up P2P media abstractions
