+++ packages/client-core/src/systems/ui/WidgetMenuView/index.tsx
@@ -35,8 +35,8 @@ import { XRState } from '@ir-engine/spatial/src/xr/XRState'
 import Icon from '@ir-engine/ui/src/primitives/mui/Icon'
 import { RegisteredWidgets, WidgetAppActions, WidgetAppState } from '../../WidgetAppService'
 
+import { MediaStreamState } from '@ir-engine/network/src/media/MediaStreamState'
 import { useMediaInstance } from '../../../common/services/MediaInstanceConnectionService'
-import { MediaStreamState } from '../../../media/MediaStreamState'
 import XRIconButton from '../../components/XRIconButton'
 import HandSVG from './back_hand_24px.svg?react'
 import styleString from './index.scss?inline'
