+++ packages/client-core/src/transports/mediasoup/MediasoupClientFunctions.ts
@@ -86,9 +86,9 @@ import {
   MediasoupTransportState,
   TransportType
 } from '@ir-engine/common/src/transports/mediasoup/MediasoupTransportState'
+import { MediaStreamState } from '@ir-engine/network/src/media/MediaStreamState'
 import { LocationInstanceState } from '../../common/services/LocationInstanceConnectionService'
 import { MediaInstanceState } from '../../common/services/MediaInstanceConnectionService'
-import { MediaStreamState } from '../../media/MediaStreamState'
 import { ChannelState } from '../../social/services/ChannelService'
 import { LocationState } from '../../social/services/LocationService'
 import { AuthState } from '../../user/services/AuthService'
