+++ packages/client-core/src/user/VideoWindows/index.tsx
@@ -31,9 +31,9 @@ import { NO_PROXY, PeerID, useMutableState } from '@ir-engine/hyperflux'
 import { NetworkState } from '@ir-engine/network'
 
 import { EngineState } from '@ir-engine/ecs'
+import { PeerMediaChannelState, PeerMediaStreamInterface } from '@ir-engine/network/src/media/PeerMediaChannelState'
 import { NetworkPeerState } from '@ir-engine/network/src/NetworkPeerState'
 import { useMediaNetwork } from '../../common/services/MediaInstanceConnectionService'
-import { PeerMediaChannelState, PeerMediaStreamInterface } from '../../media/PeerMediaChannelState'
 import { FilteredUsersState } from '../../world/FilteredUsersSystem'
 import { AuthState } from '../services/AuthService'
 import { SingleVideoWindow, SingleVideoWindowWidget } from './window'
