+++ packages/network/index.ts
@@ -44,7 +44,7 @@ export * from './src/systems/IncomingActionSystem'
 export * from './src/systems/IncomingNetworkSystem'
 export * from './src/systems/OutgoingActionSystem'
 export * from './src/systems/OutgoingNetworkSystem'
-export * from './src/webrtc/useWebRTCPeerConnection'
+export * from './src/webrtc/WebRTCPeerConnection'
 export * from './src/webrtc/WebRTCTransportFunctions'
 
 export { VideoConstants }
