+++ packages/ui/src/pages/Capture/index.tsx
@@ -33,7 +33,6 @@ import { useMediaNetwork } from '@ir-engine/client-core/src/common/services/Medi
 import { useResizableVideoCanvas } from '@ir-engine/client-core/src/hooks/useResizableVideoCanvas'
 import { useScrubbableVideo } from '@ir-engine/client-core/src/hooks/useScrubbableVideo'
 import { CaptureClientSettingsState } from '@ir-engine/client-core/src/media/CaptureClientSettingsState'
-import { MediaStreamState } from '@ir-engine/client-core/src/media/MediaStreamState'
 import { LocationState } from '@ir-engine/client-core/src/social/services/LocationService'
 import { useGet } from '@ir-engine/common'
 import {
@@ -72,6 +71,7 @@ import Canvas from '@ir-engine/ui/src/primitives/tailwind/Canvas'
 import Video from '@ir-engine/ui/src/primitives/tailwind/Video'
 
 import { SocketWebRTCClientNetwork } from '@ir-engine/client-core/src/transports/mediasoup/MediasoupClientFunctions'
+import { MediaStreamState } from '@ir-engine/network/src/media/MediaStreamState'
 import { ReferenceSpaceState } from '@ir-engine/spatial'
 import { useVideoFrameCallback } from '@ir-engine/spatial/src/common/functions/useVideoFrameCallback'
 import { Slider } from '../../../editor'
